rename `utils` to `util`
