+++ benches/benches/imf_fixdate.rs
@@ -2,7 +2,7 @@
 extern crate test;
 
 use ohkami_lib::time::UTCDateTime;
-use ohkami::utils::unix_timestamp;
+use ohkami::util::unix_timestamp;
 
 
 #[bench] fn format_imf_fixdate(b: &mut test::Bencher) {
