+++ ohkami/src/request/from_request.rs
@@ -1,5 +1,5 @@
 use std::borrow::Cow;
-use crate::{utils::ErrorMessage, IntoResponse, Request, Response};
+use crate::{util::ErrorMessage, IntoResponse, Request, Response};
 
 
 /// "Retirieved from a `Request`".
